+++ dbdaora/sync/data_sources/redis.py
@@ -1,5 +1,0 @@
-from redis import Redis
-
-
-class RedisDataSource(Redis):
-    ...